fix(pg): corregir errores de tipo y esquema en funciones PostgreSQL
Correccion masiva de 500/502 en endpoints de la API con DB_TYPE=postgres:

SQL (17 archivos):
- usp_acct, usp_ap, usp_ar, usp_doc_sales, usp_doc_purchase: agregar alias
  en subqueries COUNT para evitar column reference ambiguous
- usp_ar, usp_ap, usp_pay, usp_misc, usp_util: agregar ::VARCHAR cast en
  columnas CHAR(n) que no coinciden con RETURNS TABLE VARCHAR
- usp_cfg: cambiar BIGINT a INT en SettingId (tipo real en tabla es integer)
- usp_ops: cambiar MovementId de INT a BIGINT; agregar ::VARCHAR y ::TIMESTAMPTZ
- usp_util: agregar ::VARCHAR cast en CountryCode de TaxRetention
- sp_nomina_batch: reescribir usp_HR_Payroll_ListBatches con columnas reales
- sp_vacation_request: agregar wrapper usp_hr_vacationrequest_list (alias)
- sp_rrhh_beneficios: corregir parametros y aliases en usp_HR_ProfitSharing_List
- sp_crud_vendedores: reescribir usp_vendedores_list con columnas reales de Seller
- usp_fin_pettycash, sp_bancos_conciliacion: fixes de tipo previamente hechos

TypeScript (4 archivos):
- auditoria/service.ts: reemplazar callSpMulti(mssql) con callSp+bifurcacion
  DB_TYPE para modo PG (funciones separadas vs multi-recordset SQL Server)
- meta/service.ts: mismo patron; usar usp_Sys_Meta_TablesAndColumns_Tables
  y _Columns en PG en lugar de mssql multi-recordset
- utils/xml.ts: aceptar <rows> como contenedor XML ademas de <root>
- db/query.ts: manejar null/undefined en parametros *Xml correctamente

+++ web/api/sqlweb-pg/includes/sp/usp_cfg.sql
@@ -46,7 +46,7 @@ CREATE OR REPLACE FUNCTION usp_Cfg_AppSetting_List(
     p_company_id INT
 )
 RETURNS TABLE(
-    "SettingId" BIGINT, "Module" VARCHAR, "SettingKey" VARCHAR,
+    "SettingId" INT, "Module" VARCHAR, "SettingKey" VARCHAR,
     "SettingValue" TEXT, "ValueType" VARCHAR, "Description" VARCHAR,
     "UpdatedAt" TIMESTAMP
 )
@@ -69,7 +69,7 @@ CREATE OR REPLACE FUNCTION usp_Cfg_AppSetting_ListByModule(
     p_module     VARCHAR(60)
 )
 RETURNS TABLE(
-    "SettingId" BIGINT, "Module" VARCHAR, "SettingKey" VARCHAR,
+    "SettingId" INT, "Module" VARCHAR, "SettingKey" VARCHAR,
     "SettingValue" TEXT, "ValueType" VARCHAR, "Description" VARCHAR,
     "UpdatedAt" TIMESTAMP
 )
